MovieInfoModel DataBinding 적용
## MovieInfoModel DataBinding 적용
## BindingAdapters 파일 생성 (DataBinding에 사용 될 함수)
### 1. loadImage() : Glide를 사용하여 imageUrl를 String으로 받아와 Drawable로 교체하여 넣는 코드 수정
> - RequestOptions() : placeHolder 및 error를 설정하기 위해 생성합니다.
> - placeHolder : 리소스가 로드되는 동안 표시할 drwable resource 입니다.
> - error : 로드가 실패할 경우 표시할 drawable resource 입니다.
### 2. onSingleClick() : 중복 클릭 방지 Event Listener로 커스텀하게 만들어 사용합니다.
> interval에 들어갈 Long 값에 따라 중복 시간을 체크합니다. 중복시간 체크 기본 값은 1초 입니다. 따라서 click을 누르고 1초안에 계속 클릭한다면 다음 Click Listener 는 동작하지 않습니다.

+++ app/src/main/java/com/example/moviesearchapp/data/mapper/ObjectMapper.kt
@@ -13,7 +13,7 @@ object ObjectMapper {
             pubDate= it.pubDate,
             director= it.director,
             actor= it.actor,
-            userRating= it.userRating
+            userRating= it.userRating.toFloat()
         )
     }
 
@@ -25,6 +25,6 @@ object ObjectMapper {
         pubDate= this.pubDate,
         director= this.director,
         actor= this.actor,
-        userRating= this.userRating
+        userRating= this.userRating.toFloat()
     )
 }
